feat(web-components): allow custom signer instead of keplr
